mockapi/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
sqlalchemy==2.0.36
sqlite3
openai==1.56.2
aiosqlite==0.20.0
asyncio
typing-extensions==4.12.2